Run user process in user-mode (CPL=3).
This hit a bunch of roadblocks but I think it's finally done.
(I've been staring at this all day - I'm not confident I
remember the order in which things happened, lol).

First, simply OR-ing the segment selectors specified in the user TSS
with 0x3 caused isr10 to fire. I investigated this for the longest time :(.
I think it helped here to note the difference between steps 9 and 14
in section 7.3 of the ia32 manual:

Step 9 says the segment selectors are loaded.
Step 14 says the segment descriptors (based on the loaded segment selectors)
are validated and loaded.

This was important for me because I noticed (via Qemu's `info registers`)
that some of the segment registers would be loaded (usually all except cs)
but the descriptor(s) from the GDT would not be loaded. So it gave me an idea
where something was going wrong.

I think I then realized I was simply, for example, OR-ing the cs register with
0x3 - so it was still pointing to a GDT entry with a DPL of 0 (:O). So I decided
I'd need separate GDT entries for the USER task.

So, I set up entries 3 and 4 in the GDT (which were previously just 0). These
entries were similar to the kernel entries except their DPL was 3 and not 0.
This was also a challenge for a while: It seemed that making the entries exactly
identical to the kernel entries worked fine but setting their DPL to 3 caused
"Invalid TSS Exception." I thought this was weird because what kind of
restrictions would not allow access to a descriptor with a DPL of 3? Well, after
hours of staring at the manual and surfing the web, I landed at
https://forum.osdev.org/viewtopic.php?f=1&t=610 which gave me the idea of OR-ing
the TSS's segment selectors with 0x3 again. I thought it was strange but for
whatever reason, that worked! I would have thought segment selectors with higher
privilege levels would be less likely be an issue.

Anyway, the next challenge was quite unexpected but very welcome: Page fault!
It was pretty weird. It appeared that the attempt to fetch the first
instruction of the task (at 0x30000000) cause a page fault with an error code of
5. I thought this really strange as the code set the lowest 3 bits for the page
table entry for 0x30000000 to 0x7. I confirmed this by inspecting memory (page
table content) via gdb and QEMU so I was pretty confused.

So I started reading up on page fault semantics. The error code 5 meant the entry
was present but access was violated. I came across section 4.6.1 which had the
quote "If the U/S flag (bit 2) is 0 in at least one of the paging-structure
entries, the address is a supervisor-mode address." so I went back and looked at
the page directory structure and indeed the relevant entry was OR-ed with 0x3 but
I needed to OR it with 0x7 so the U/S bit would be set.

After that, I ran into a General Protection fault with error code 8. (I should
note that when facing the page fault problem, I observed that the page directory
entry would be marked accessed but not the page table entry but with the #GP
error I observed that they were both accessed so I knew it was progress.) The #GP
error was easy to fix: I had left the app with the instruction: `jmp 0xb:exit`
(0xb = 0x8 | 0x3) but since I was now using a different code segment (0x18) I
needed to use `jmp 0x1b:exit` instead and indeed this fixed it :).