Memory Management - A First Stab
Tried to implement a buddy allocator. It's real simple:

The region of physical memory where dynamic memory allocations can come
is defined
	[PAGE_ALIGN_UP(&_bss_end), SYSTEM_RAM_SIZE)
This memory is split up into MAX_ORDER (currently 8) "zones".
Each zone consists of "mem_blocks". A mem_block in a given zone consists
consists of a number of pages. This number of pages is 2^zone->order.

That's the basic idea.

We try to do splitting and coalescing. So, if a block cannot fulfill an
allocation (because its freelist is empty), we try to borrow from the
"next-higher-order zone", NHOZ. A block from the NHOZ is split up into 2
child blocks. One child block is added to the borrowing zone's freelist
and the order is returned to fulfill the allocation. When one of the
child block's is being freed, we try to coalesce it with its buddy (the
other child block) and return the the resulting block to the NHOZ. (We
know that a block should be coalesced with its buddy if its order is not
its "trueorder"). Of course, we only do this coalescing if the block's
buddy is also FREE.

That's it, really. Quite Simple.
Signed-off-by: dafidi <awogbemiladavid@gmail.com>