Giving filesystem a real go: some foundational stuff.
The filesystem design can be seen in fs_params.md and
templates/filesystem/readme.asm. It tries to use the same idea as the
EXT filesystems where the key structure is the fnode which contains
some metadata fields and an array of pointers to sectors. EXT uses
direct pointers, singly indirect pointers and doubly indirect
pointers. This commit only uses direct pointers for simplicity at
introduction - later indirect pointers will be added.

The design specifies 5 "components" of the filesystem:

1. master_record
2. fnode_bitmap
3. sector_bitmap
4. fnode_table
5. data_blocks

master_record

The master_record occupied sector zero on the disk. It is defined as
the struct

struct fs_master_record {
    struct fnode_location_t root_dir_fnode_location;
    uint32_t fnode_bitmap_start_sector;
    uint32_t fnode_bitmap_size;
    uint32_t fnode_table_start_sector;
    uint32_t sector_bitmap_start_sector;
    uint32_t sector_bitmap_size;
    uint32_t data_blocks_start_sector;
};

whose fields have descriptive names.
From the master_record, the kernel can figure our where the fnode for
the root directory is and the size and the location of other components on the disk.

fnode_bitmap

This is a bitmap of fnodes where 0 indicates that an fnode is free/unused and 1
indicates that it is used. The filesystem has a fixed number of fnodes so the
fnode_bitmap size is also fixed.

sector_bitmap

This is a bitmap similar to the fnode_bitmap but tracing sectors not fnodes.
The sector_bitmap is fixed in size since the disk should(?) be fixed in size.

fnode_table

This is an array of fnodes. It can be considered to be the first portion of the
"data" section of the filesystem. The previous 3 sections can be considered
"metadata" sections.

data_blocks

This is raw filesystem content which are accessed via sector indexes in fnodes.
Broadly, there are 3 types of content:

i)   struct dir_info - This can only be found in the content of a directory
	fnode. It describes the directory.
ii)  struct dir_entry - This can only be found in the content of a directory
	fnode. It describes a file or folder within the containing directory.
iii) raw file content - This is the content of a file, e.g. text, binary,
	image, etc.