Implementation: Dynamic Kernel Memory Object Allocation
This change adds support for dynamically allocating and freeing kernel,
think kmalloc/kfree. It relies on the previously implemented Dynamic
Zone Allocation.

It's a pretty simple concept: keep a "cache" of "objects" of a range of
sizes, 32, 64, 128, 256, 512, 1024, and 2048 in this case. (Objects can
be thought of as buffers.)
Each cache maintains a list of free objects and a list of used objects.
Allocation and de-allocation is basically moving object across these
two lists.

It's worth noting that one thing that isn't taking into account at all
is synchronization. The code assumes a uniprocessor (UP) environment.
Support for SMP is a stretch goal :)