Fixup: Correct a few minor errors introduced by the Dynamic Object Allocation patch
This patch expanded the object allocation test a little. This caused
the kernel size to be big enough that the kernel was overwriting the
bootloader and causing hard-to-understand problems. A simple fix is
to load the kernel above the bootloader (i.e 0x8400). This gives the
kernel room to grow up to 0xb8000, ~700KB which should be plenty of
room.

The errors correct were:

- The object_block_ptr field should have type mem_block, not
memory_block.
- The skip_size calculation in memory_object_cache_init has the wrong
expression on the RHS - the shift operation should apply only to the 1.