Implementation: Add functionality to create a new file in the fs
This change makes is possible to create new file via the "newf"
shell command. The process of creating a file is actually quite
straightforward, but two things make this such a big patch:
error-handling and the fact that the bitmaps are currently on-disk.
The latter makes the former more complicated. Then again, it might
still be just as complicated when the bitmaps are in-memory because
they'd still have to be written to disk and kept in sync.

A couple of minor changes to note:

- struct dir_info now includes the directory's name - this means that
we can figure out an fnode's name using just the fnode. This should
also be done for regular files.