Adjustment: Use larger buffers in query_free_{sectors, fnodes}
This patch wanted to improve the performance of
query_free_{sectors, fnodes} by reading in the bitmaps in larger
chunks - but that didn't really help.
Still it's better to read them in larger chunks so the change is
staying.

Other minor changes:

- Adds a macro to time functions. The mark time function is kinda faulty
it relies on a variable that is incremented during interrupts so it is not so
useful on functions that disable interrupts.

- Adds an implementation of the ATA "IDENTIFY DEVICE" command execution.

- Refactors read_from_storage_disk so that we can break up read requests
to sizes smaller than the device's DRQ DATA BLOCK setting.