Adjustment: Skip metadata components in query_free_sectors scan
This patch changes the search space of query_free_sectors so that
we start searching for a free sector *after* the sectors occupied by the
filesystem's metadata components (which will always not be free).
Sometimes, we should just let occam's razor run its course.

That being said, while it might be a micro-optimization, the actual
bit-checking logic (get_bit) can probably be improved.

Other minor changes:

- Introduce bit_scan_forward implementation. This is not actually used,
but may be useful in the future.