Adjustment: Clean up shell.c code
This patch tries to clean up the very ugly-loking shell.c code.
Now, it isn't perfect, but it's better.

Still, the handling of input keystrokes could be improved upon,
perhaps with the use of a circular buffer. This would actually
allow keystrokes to still be accepted/stored while a command is
being processed.