Implementation: Add minimal change-directory functionality
This change adds support for changing the working directory of the shell
to a subdirectory of the current working directory. It cannot yet use
names with slashes or interpret multi-level subdirectories. It also
cannot change-directory into a parent dircetory. These will all still
require a substantial amount of work but this is a good start still.