Implementation: x86_64 64-bit Support
This adds support for running the kernel in IA32e mode, 64-bit
mode specifically and defaults to building and running in 64-bit
mode while keeping current functionality working for 32-bit mode.
Enabling 64-bit mode required support for 4-level paging which has
been added in kernel/asm/kernel64.asm.
64-bit mode optionally uses Interrupt Stack Tables (ISTs) to handle
interrupts. The IDT is initialized with entrys that have a 3-bit
field which represents an index into the IST section of the TSS.
If an IDT entry has o in its IST field, it does not use IDT but a
modified version of the legacy interrupt handling mechanism. If it
has a non-zero IST, its IST field value points to one of the 7
64-bit ISTs in the TSS. Each IST in the TSS points to a location in
memory and the stack pointer is loaded with this memory location when
an interrupt associated with that IDT entry fires.
This patch allocates 7 4KiB regions to serve as interrupt stacks. This
patch sets the IST field is set to zero for all interrupts excepts:

- isr 13 (General Protection Fault)
- isr 14 (Page Fault)
- irq 0  (isr 32) (Timer Interrupt)
- irq 1  (isr 33) (Keyboard Interrupt)
- irq 4  (isr 36) (Not sure what device this interrupt is servicing)
- irq 14 (isr 46) (Disk Interrupt)

Ths isrs/irqs for which a non-zero IST has been set have been picked
based on what interrupts have actually been firing. As the O/S
develops, more interrupts may fire and need to have their own stack
set up.
Leaving all the IST entries as zero should have worked but did not
(Double/Triple Fault kept firing).
This was the motivation for using the IST mechanism.

Minor changes:
 - Getting rid of gcc warnings for both 32- and 64- bit builds.
 - Switching to protected mode before loading the kernel in
   boot_sect_2.asm.
 - Improving boot/disk_load.asm to read in batches of 128 sectors.