Implementation: Allocate memory order zones based on available memory
The main objective of this patch is to make the memory management system
dynamically determine the size and the regions of available memory and
allocate order zones based on that.
It turns out detecting available memory is a little trickier than one
would expect. The memory map returns by the bios via the e820 command is
useful but not prefect. It has been the case that areas of memory
reported by this map as available are not writeable, even though they
don't overlap with areas not reported as available. So, a little bit of
hacking has been added to cater to some of the observed peculiarities such
as:
 - when we have 4GiB of ram available, we receive this map of available
   regions:
	[1MiB, 3GiB) U [4GiB, 5GiB)
   but it turns out that we can only rely on being able to write to:
	[1MiB, 3GiB) U [4GiB, 4GiB + 0xb000)

The order of items in init() has also been changed so that interrupts don't
fire before sttart_interrupts is called (this hasn't really been a problem
but seems like the right thing to do).